requirements.txt
tornado>=2.1.1
sockjs-tornado>=1.0.1
pika>=0.9.14
